Initial Route Refactoring
